% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_segment_overlap.R
\name{get_gene_seg_overlap}
\alias{get_gene_seg_overlap}
\title{Get Gene Overlaps with CNVkit segments}
\usage{
get_gene_seg_overlap(segs.df, gene.annot.df, overlap.prop = 0.5)
}
\arguments{
\item{segs.df}{data.frame of the CNVkit segment file (.cns).}

\item{gene.annot.df}{data.frame of the gene annotations.}

\item{overlap.prop}{Proportion that the segment must overlap with
a gene in order to be considered overlapping. The smaller the value the 
more focal the copy number alteration you want to associate with gene}
}
\value{
data.frame that contains the overlap of the input gene annotation
 with the input segment data.
}
\description{
Get Gene Overlaps with CNVkit segments
}

